Removing IiifPrintHelperBehavior from generator
The `app/helpers/iiif_print_helper.rb` is loaded because `app/helpers`
in an engine are loaded into the downstream application.

We don't need to install a module that includes another module which is
defined in the application.

This moves in the direction of minimizing what the IiifPrint engine
generates into downstream applications.
