--- lib/generators/iiif_print/templates/iiif_print_helper.rb
@@ -1,3 +1,0 @@
-module IiifPrintHelper
-  include IiifPrint::IiifPrintHelperBehavior
-end
